+++ src/lib/meta.js
@@ -1,5 +1,6 @@
 var geojsonRandom = require('geojson-random'),
-    geojsonExtent = require('geojson-extent');
+    geojsonExtent = require('geojson-extent'),
+    geojsonFlatten = require('geojson-flatten');
 
 module.exports.clear = function(context) {
     context.data.clear();
@@ -11,4 +12,8 @@ module.exports.random = function(context, count, type) {
 
 module.exports.bboxify = function(context) {
     context.data.set({ map: geojsonExtent.bboxify(context.data.get('map')) });
+};
+
+module.exports.flatten = function(context) {
+    context.data.set({ map: geojsonFlatten(context.data.get('map')) });
 };